Move contentloader to a module
This is commit 1of2 to reduce memory use of the contentloader.
Often the contentloader will be invoked a dozen times and eleven
of those processes will sit waiting for the process lock, already
having imported a lot of modules and using a lot of memory while
idle. By placing the process lock *before* the imports, memory
use is greatly reduced.